Adding SeparableBinaryMorphologyImageFilter base class
This class in a combination of three things, the
RecursiveSeparableImageFilter, the BinaryMorphologyImageFilter with
the addtion of a new way to traverse the data in the three directions
in one update, and filtering the image as arrays in each direction.

The binary mophology algorithms that this class is designed to work
with only works for box shaped struturing elements.